@@ -6,7 +6,7 @@ description = "Command line interface for analyzing Python imports."
 authors = [
     {name = "David Seddon", email = "david@seddonym.me"},
 ]
-requires-python = ">=3.9"
+requires-python = ">=3.10"
 dependencies = [
     "click>=6",
     "grimp==3.13",
@@ -20,7 +20,6 @@ classifiers = [
     "Operating System :: POSIX",
     "Operating System :: Microsoft :: Windows",
     "Programming Language :: Python",
-    "Programming Language :: Python :: 3.9",
     "Programming Language :: Python :: 3.10",
     "Programming Language :: Python :: 3.11",
     "Programming Language :: Python :: 3.12",
@@ -84,6 +83,7 @@ import-linter = { workspace = true }
 dev = [
     "mypy>=1.18.2",
     "pre-commit>=4.3.0",
+    "pyupgrade>=3.21.0",
     "ruff>=0.13.2",
 ]
 docs = [
